refactor: отредактирован readme
