+++ src/services/actions/reset-password.jsx
@@ -4,7 +4,7 @@ import {
   RESET_PASSWORD_REQUEST,
   RESET_PASSWORD_SUCCESS,
   RESET_PASSWORD_FAILED,
-} from '../../utils/action-types';
+} from './action-types';
 
 export const resetPassword = (form) => (dispatch) => {
   dispatch({ type: RESET_PASSWORD_REQUEST });
